[Sharktank][Llama][FP8] Minimal changes for numerically correct fp8 (#859)
This patch enables the use of quark quantized models of the latest
generation. Many changes were required to enable parity with the source
model which is very sensitive to any numerical fluctuations. There is a
test added to maintain this parity but is disabled until I can get the
relevant data set up on the ci machine.

+++ sharktank/sharktank/layers/linear.py
@@ -78,7 +78,6 @@ class LinearLayer(ThetaLayer):
             x = qdq_input.quantize(x).unpack().dequant()
 
         y = ops.linear(x, weight, bias)
-
         # Unconditionally dequantize.
         if isinstance(y, QuantizedTensor):
             y = y.unpack().dequant()
@@ -88,7 +87,7 @@ class LinearLayer(ThetaLayer):
         # level to do this, but for now its here.
         if not isinstance(y, QuantizedTensor):
             if y.dtype == torch.float8_e4m3fnuz:
-                y = ops.to(y, torch.float16)
+                y = ops.to(y, torch.bfloat16)
                 return y
         if qdq_output is not None:
             y = qdq_output.quantize(y).unpack().dequant()
